pal: linux: split implementation into multiple files
