+++ src/pal/linux/CMakeLists.txt
@@ -1,0 +1,9 @@
+target_sources(pal PRIVATE p-alert.c)
+target_sources(pal PRIVATE p-id.c)
+target_sources(pal PRIVATE p-init.c)
+target_sources(pal PRIVATE p-shell.c)
+target_sources(pal PRIVATE p-state.c)
+target_sources(pal PRIVATE p-time.c)
+target_sources(pal PRIVATE p-volume.c)
+
+target_sources(pal PRIVATE l-info.c)
