+++ src/pal/linux/impl.h
@@ -1,0 +1,7 @@
+#ifndef _PAL_LINUX_IMPL_H
+#define _PAL_LINUX_IMPL_H
+
+extern bool linux_has_beepemu;
+extern long linux_start_msec;
+
+#endif // _PAL_LINUX_IMPL_H
