+++ src/pal/linux/p-time.c
@@ -1,0 +1,32 @@
+#include <time.h>
+#include <unistd.h>
+
+#include <pal.h>
+
+#include "impl.h"
+
+long linux_start_msec;
+
+uint32_t
+pal_get_counter(void)
+{
+    struct timespec time;
+    clock_gettime(CLOCK_BOOTTIME, &time);
+
+    long time_msec = time.tv_sec * 1000 + time.tv_nsec / 1000000;
+    return time_msec - linux_start_msec;
+}
+
+uint32_t
+pal_get_ticks(unsigned ms)
+{
+    return ms;
+}
+
+void
+pal_sleep(unsigned ms)
+{
+    LOG("entry, ms: %u", ms);
+
+    usleep(ms);
+}
